# services/frontend/openapi.json
{
  "components": {
    "schemas": {
      "AiTestRequest": {
        "description": "Credentials as currently typed, so a key can be checked before saving.",
        "properties": {
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          },
          "base_url": {
            "title": "Base Url",
            "type": "string"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "timeout_seconds": {
            "default": 10.0,
            "minimum": 1.0,
            "title": "Timeout Seconds",
            "type": "number"
          }
        },
        "required": [
          "base_url",
          "model"
        ],
        "title": "AiTestRequest",
        "type": "object"
      },
      "AiTestResult": {
        "properties": {
          "latency_ms": {
            "title": "Latency Ms",
            "type": "integer"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok",
          "message",
          "latency_ms"
        ],
        "title": "AiTestResult",
        "type": "object"
      },
      "ClearedModel": {
        "properties": {
          "deleted": {
            "title": "Deleted",
            "type": "integer"
          }
        },
        "required": [
          "deleted"
        ],
        "title": "ClearedModel",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HistoryPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/OperationModel"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "HistoryPage",
        "type": "object"
      },
...
      "OperationModel": {
        "properties": {
          "added_tracks": {
            "items": {
              "$ref": "#/components/schemas/TrackModel"
            },
            "title": "Added Tracks",
            "type": "array"
          },
          "app": {
            "enum": [
              "radarr",
              "sonarr"
            ],
            "title": "App",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "destination_path": {
            "title": "Destination Path",
            "type": "string"
          },
          "dry_run": {
            "default": false,
            "title": "Dry Run",
            "type": "boolean"
          },
          "duration_ms": {
            "default": 0,
            "title": "Duration Ms",
            "type": "integer"
          },
          "episodes": {
            "items": {
              "type": "integer"
            },
            "title": "Episodes",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "log": {
            "items": {
              "$ref": "#/components/schemas/LogEntryModel"
            },
            "title": "Log",
            "type": "array"
          },
          "media_file": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Media File"
          },
          "move_status": {
            "enum": [
              "DeferMove",
              "MoveComplete",
              "RenameRequested"
            ],
            "title": "Move Status",
            "type": "string"
          },
          "output_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Bytes"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "rejected_tracks": {
            "items": {
              "$ref": "#/components/schemas/RejectedTrackModel"
            },
            "title": "Rejected Tracks",
            "type": "array"
          },
          "removed_tracks": {
            "items": {
              "$ref": "#/components/schemas/RemovedTrackModel"
            },
            "title": "Removed Tracks",
            "type": "array"
          },
          "season": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Season"
          },
          "source_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Bytes"
          },
          "source_path": {
            "title": "Source Path",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "transfer_mode": {
            "default": "",
            "title": "Transfer Mode",
            "type": "string"
          }
        },
        "required": [
          "id",
          "created_at",
          "app",
          "title",
          "move_status",
          "reason",
          "source_path",
          "destination_path",
          "media_file",
          "transfer_mode",
          "season",
          "episodes",
          "added_tracks",
          "rejected_tracks",
          "removed_tracks",
          "log",
          "duration_ms",
          "source_bytes",
          "output_bytes",
          "dry_run"
        ],
        "title": "OperationModel",
        "type": "object"
      },
...
      "SettingsPatch": {
        "additionalProperties": false,
        "description": "Only the fields actually sent are changed.",
        "properties": {
          "ai_api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Api Key"
          },
          "ai_base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Base Url"
          },
          "ai_max_entries": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Max Entries"
          },
          "ai_mode": {
            "anyOf": [
              {
                "enum": [
                  "off",
                  "fallback",
                  "always",
                  "verify"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Mode"
          },
          "ai_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Model"
          },
          "ai_name_tracks": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Name Tracks"
          },
          "ai_timeout_seconds": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Timeout Seconds"
          },
          "dedupe": {
            "anyOf": [
              {
                "enum": [
                  "off",
                  "language",
                  "language_codec"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dedupe"
          },
          "free_space_factor": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Free Space Factor"
          },
          "history_max_records": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "History Max Records"
          },
          "job_ttl_seconds": {
            "anyOf": [
              {
                "minimum": 60.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Ttl Seconds"
          },
          "keep_audio_languages": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Keep Audio Languages"
          },
          "keep_subtitle_languages": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Keep Subtitle Languages"
          },
          "log_level": {
            "anyOf": [
              {
                "enum": [
                  "TRACE",
                  "DEBUG",
                  "INFO",
                  "SUCCESS",
                  "WARNING",
                  "ERROR",
                  "CRITICAL"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Log Level"
          },
          "max_concurrent_muxes": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Concurrent Muxes"
          },
          "max_external_tracks": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max External Tracks"
          },
          "mux_timeout_seconds": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mux Timeout Seconds"
          },
          "operation_log_max_entries": {
            "anyOf": [
              {
                "maximum": 5000.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operation Log Max Entries"
          },
          "preserve_ownership": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preserve Ownership"
          },
          "skip_image_subtitles": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skip Image Subtitles"
          },
          "skip_undetermined_language": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skip Undetermined Language"
          },
          "sub_charset": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sub Charset"
          }
        },
        "title": "SettingsPatch",
        "type": "object"
      },
      "SettingsView": {
        "properties": {
          "ai_api_key_set": {
            "title": "Ai Api Key Set",
            "type": "boolean"
          },
          "ai_base_url": {
            "title": "Ai Base Url",
            "type": "string"
          },
          "ai_max_entries": {
            "title": "Ai Max Entries",
            "type": "integer"
          },
          "ai_mode": {
            "enum": [
              "off",
              "fallback",
              "always",
              "verify"
            ],
            "title": "Ai Mode",
            "type": "string"
          },
          "ai_model": {
            "title": "Ai Model",
            "type": "string"
          },
          "ai_name_tracks": {
            "title": "Ai Name Tracks",
            "type": "boolean"
          },
          "ai_timeout_seconds": {
            "title": "Ai Timeout Seconds",
            "type": "number"
          },
          "dedupe": {
            "enum": [
              "off",
              "language",
              "language_codec"
            ],
            "title": "Dedupe",
            "type": "string"
          },
          "free_space_factor": {
            "title": "Free Space Factor",
            "type": "number"
          },
          "history_max_records": {
            "title": "History Max Records",
            "type": "integer"
          },
          "job_ttl_seconds": {
            "title": "Job Ttl Seconds",
            "type": "number"
          },
          "keep_audio_languages": {
            "items": {
              "type": "string"
            },
            "title": "Keep Audio Languages",
            "type": "array"
          },
          "keep_subtitle_languages": {
            "items": {
              "type": "string"
            },
            "title": "Keep Subtitle Languages",
            "type": "array"
          },
          "locked": {
            "items": {
              "type": "string"
            },
            "title": "Locked",
            "type": "array"
          },
          "log_level": {
            "title": "Log Level",
            "type": "string"
          },
          "max_concurrent_muxes": {
            "title": "Max Concurrent Muxes",
            "type": "integer"
          },
          "max_external_tracks": {
            "title": "Max External Tracks",
            "type": "integer"
          },
          "mux_timeout_seconds": {
            "title": "Mux Timeout Seconds",
            "type": "number"
          },
          "operation_log_max_entries": {
            "title": "Operation Log Max Entries",
            "type": "integer"
          },
          "preserve_ownership": {
            "title": "Preserve Ownership",
            "type": "boolean"
          },
          "skip_image_subtitles": {
            "title": "Skip Image Subtitles",
            "type": "boolean"
          },
          "skip_undetermined_language": {
            "title": "Skip Undetermined Language",
            "type": "boolean"
          },
          "sub_charset": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sub Charset"
          }
        },
        "required": [
          "dedupe",
          "skip_image_subtitles",
          "skip_undetermined_language",
          "max_external_tracks",
          "sub_charset",
          "keep_audio_languages",
          "keep_subtitle_languages",
          "mux_timeout_seconds",
          "free_space_factor",
          "preserve_ownership",
          "max_concurrent_muxes",
          "job_ttl_seconds",
          "history_max_records",
          "operation_log_max_entries",
          "ai_mode",
          "ai_base_url",
          "ai_model",
          "ai_timeout_seconds",
          "ai_max_entries",
          "ai_name_tracks",
          "ai_api_key_set",
          "log_level",
          "locked"
        ],
        "title": "SettingsView",
        "type": "object"
      },
      "StatsModel": {
        "properties": {
          "deferred": {
            "title": "Deferred",
            "type": "integer"
          },
          "last_24h": {
            "title": "Last 24H",
            "type": "integer"
          },
          "muxed": {
            "title": "Muxed",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "tracks_added": {
            "title": "Tracks Added",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "muxed",
          "deferred",
          "tracks_added",
          "last_24h"
        ],
        "title": "StatsModel",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "muxarr",
    "version": "0.9.2"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/history": {
      "delete": {
        "operationId": "clear_history_v1_history_delete",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClearedModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Clear History",
        "tags": [
          "history"
        ]
      },
      "get": {
        "operationId": "list_history_v1_history_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "move_status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Move Status"
            }
          },
          {
            "in": "query",
            "name": "app",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "App"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List History",
        "tags": [
          "history"
        ]
      }
    },
    "/v1/history/{operation_id}": {
      "get": {
        "operationId": "get_operation_v1_history__operation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "operation_id",
            "required": true,
            "schema": {
              "title": "Operation Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Operation",
        "tags": [
          "history"
        ]
      }
    },
...
    "/v1/settings": {
      "get": {
        "operationId": "read_settings_v1_settings_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Settings",
        "tags": [
          "settings"
        ]
      },
      "patch": {
        "operationId": "update_settings_v1_settings_patch",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Settings",
        "tags": [
          "settings"
        ]
      }
    },
    "/v1/settings/ai/test": {
      "post": {
        "operationId": "probe_ai_provider_v1_settings_ai_test_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AiTestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiTestResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Probe Ai Provider",
        "tags": [
          "settings"
        ]
      }
    },
    "/v1/stats": {
      "get": {
        "operationId": "get_stats_v1_stats_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Stats",
        "tags": [
          "history"
        ]
      }
    },
...
  }
}